Add mixnode_packager.sh
This commit introduced the "mixnode_packager.sh" which was designed to
handle and cordinate the automation of various activities involved in
running a Nym Mix Node. The "mixnode_packager.sh" was also designed to
be interactive with prospective users and to serve as an Inter-Process
Communication trigger in the scheme of technical operations.